Workaround Android 4.4.2 browser not sending keypress events
Bug report: https://code.google.com/p/android/issues/detail?id=68284

Basically we rely on input event instead when possible.

It mean that we will reformat the expiry field slighly more often that
really needded, but I don't think it matter that much.

We could use a similar approach for the CC number field, but in the end
since it only add spaces, a broken formatting is not as bad as on expiry.