Fix single digit expiry > 2 with leading zero.
The expiry formatter looks for two-digit leading numbers with the second digit
great than 2 to detect non-months (e.g. '1' followed by '3' is formatted as
'01 / 3'). However, that did not account for leading zeroes, which was causing
'0' followed by a digit greater than 2 to reformat incorrectly (e.g. '00 / 5').
This change explicitly looks for leading zeroes.