Fix build, bump deps for node4, sync currentTarget.
Existing `master` was failing. We specify our jquery dependency with
`>= 1.5` which at some point between the previous PR and now started
picking up a `>2.0.0` release. jQuery2 doesn't work with the old version
of jsdom we use. The latest version of jsdom requires nodejs >= 4.0.

The latest jsdom also introduced some issues with our tests: previously,
`selectionStart` was undefined, but they have since added support for
it. Some tests set values directly e.g. with `.val('4242')` and then try
to format by triggering an event. This once worked when `selectionStart`
was `undefined`, but since jsdom added a `selectionStart` API, our
`formatCardNumber` implementation saw it as `0` (rather than
`undefined`), not matching the end of the input, and returned early.
Those tests were updated to move `selectionStart` to the correct
position before triggering the formatting event.

The latest jsdom + zepto also results in `e.currentTarget` being
undefined in async handling of events, for some reason. In the `format*`
funcs, `e.currentTarget` is saved to a local synchronously. In the
`reFormat*` funcs, it was grabbed async, causing some tests to
fail. The `reFormat*` funcs were updated to grab `e.currentTarget`
synchronously. This was the only change to production code.